backend/requirements.txt
Django
psycopg[binary]
django-ninja
Pillow
pywebpush
